+++ SMLHelper/Patchers/CraftDataPatcher_Subnautica.cs
@@ -1,11 +1,16 @@
 ﻿#if SUBNAUTICA
 namespace SMLHelper.V2.Patchers
 {
+    using Harmony;
+    using SMLHelper.V2.Assets;
+    using SMLHelper.V2.Handlers;
+    using System;
     using System.Collections.Generic;
 
     internal partial class CraftDataPatcher
     {
         internal static IDictionary<TechType, ITechData> CustomTechData = new SelfCheckingDictionary<TechType, ITechData>("CustomTechData", AsStringFunction);
+        internal static IDictionary<TechType, ITechData> PatchedTechData = new SelfCheckingDictionary<TechType, ITechData>("CustomTechData", AsStringFunction);
         internal static IDictionary<TechType, TechType> CustomHarvestOutputList = new SelfCheckingDictionary<TechType, TechType>("CustomHarvestOutputList", AsStringFunction);
         internal static IDictionary<TechType, HarvestType> CustomHarvestTypeList = new SelfCheckingDictionary<TechType, HarvestType>("CustomHarvestTypeList", AsStringFunction);
         internal static IDictionary<TechType, int> CustomFinalCutBonusList = new SelfCheckingDictionary<TechType, int>("CustomFinalCutBonusList", TechTypeExtensions.sTechTypeComparer, AsStringFunction);
@@ -22,8 +27,22 @@ namespace SMLHelper.V2.Patchers
             CustomTechData.Add(techType, techData);
         }
 
-        private static void PatchForSubnautica()
+        private static void PatchForSubnautica(HarmonyInstance harmony)
         {
+            IngameMenuHandler.Main.RegisterOnQuitEvent(() => CustomTechData = PatchedTechData);
+
+            harmony.Patch(AccessTools.Method(typeof(CraftData), nameof(CraftData.PreparePrefabIDCache)),
+               postfix: new HarmonyMethod(AccessTools.Method(typeof(CraftDataPatcher), nameof(CraftDataCachePostfix))));
+            harmony.Patch(AccessTools.Method(typeof(CraftData), nameof(CraftData.PreparePrefabIDCache)),
+               postfix: new HarmonyMethod(AccessTools.Method(typeof(CraftDataPatcher), nameof(PreparePrefabIDCachePostfix))));
+
+            harmony.Patch(AccessTools.Method(typeof(CraftData), nameof(CraftData.Get)),
+                postfix: new HarmonyMethod(AccessTools.Method(typeof(CraftDataPatcher), nameof(CraftDataCachePostfix))));
+            harmony.Patch(AccessTools.Method(typeof(CraftData), nameof(CraftData.Get)),
+               postfix: new HarmonyMethod(AccessTools.Method(typeof(CraftDataPatcher), nameof(PreparePrefabIDCachePostfix))));
+
+
+
             // Direct access to private fields made possible by https://github.com/CabbageCrow/AssemblyPublicizer/
             // See README.md for details.
             PatchUtils.PatchDictionary(CraftData.harvestOutputList, CustomHarvestOutputList);
@@ -37,7 +56,12 @@ namespace SMLHelper.V2.Patchers
             PatchUtils.PatchDictionary(CraftData.backgroundTypes, CustomBackgroundTypes);
             PatchUtils.PatchList(CraftData.buildables, CustomBuildables);
 
-            AddCustomTechDataToOriginalDictionary();
+        }
+
+        private static void CraftDataCachePostfix()
+        {
+            if(CustomTechData.Count > 0)
+                AddCustomTechDataToOriginalDictionary();
         }
 
         private static void AddCustomTechDataToOriginalDictionary()
@@ -46,59 +70,73 @@ namespace SMLHelper.V2.Patchers
             short replaced = 0;
             foreach (TechType techType in CustomTechData.Keys)
             {
+                bool techDataExists = CraftData.techData.ContainsKey(techType);
                 ITechData smlTechData = CustomTechData[techType];
 
-                var techDataInstance = new CraftData.TechData
+                if (techDataExists && CraftData.Get(techType) != smlTechData)
                 {
-                    _techType = techType,
-                    _craftAmount = smlTechData.craftAmount
-                };
-
-                var ingredientsList = new CraftData.Ingredients();
-
-                if (smlTechData.ingredientCount > 0)
-                {
-                    for (int i = 0; i < smlTechData.ingredientCount; i++)
+                    var techDataInstance = new CraftData.TechData
                     {
-                        IIngredient smlIngredient = smlTechData.GetIngredient(i);
+                        _techType = techType,
+                        _craftAmount = smlTechData.craftAmount
+                    };
 
-                        var ingredient = new CraftData.Ingredient(smlIngredient.techType, smlIngredient.amount);
-                        ingredientsList.Add(smlIngredient.techType, smlIngredient.amount);
-                    }
-                    techDataInstance._ingredients = ingredientsList;
-                }
+                    var ingredientsList = new CraftData.Ingredients();
 
-                if (smlTechData.linkedItemCount > 0)
-                {
-                    var linkedItems = new List<TechType>();
-                    for (int l = 0; l < smlTechData.linkedItemCount; l++)
+                    if (smlTechData.ingredientCount > 0)
                     {
-                        linkedItems.Add(smlTechData.GetLinkedItem(l));
+                        for (int i = 0; i < smlTechData.ingredientCount; i++)
+                        {
+                            IIngredient smlIngredient = smlTechData.GetIngredient(i);
+
+                            var ingredient = new CraftData.Ingredient(smlIngredient.techType, smlIngredient.amount);
+                            ingredientsList.Add(smlIngredient.techType, smlIngredient.amount);
+                        }
+                        techDataInstance._ingredients = ingredientsList;
                     }
-                    techDataInstance._linkedItems = linkedItems;
-                }
 
-                bool techDataExists = CraftData.techData.ContainsKey(techType);
+                    if (smlTechData.linkedItemCount > 0)
+                    {
+                        var linkedItems = new List<TechType>();
+                        for (int l = 0; l < smlTechData.linkedItemCount; l++)
+                        {
+                            linkedItems.Add(smlTechData.GetLinkedItem(l));
+                        }
+                        techDataInstance._linkedItems = linkedItems;
+                    }
 
-                if (techDataExists)
-                {
-                    CraftData.techData.Remove(techType);
-                    Logger.Log($"{techType} TechType already existed in the CraftData.techData dictionary. Original value was replaced.", LogLevel.Warn);
-                    replaced++;
+                    if (techDataExists)
+                    {
+                        CraftData.techData.Remove(techType);
+                        Logger.Log($"{techType} TechType already existed in the CraftData.techData dictionary. Original value was replaced.", LogLevel.Warn);
+                        replaced++;
+                    }
+                    else
+                    {
+                        added++;
+                    }
+                    CraftData.techData.Add(techType, techDataInstance);
+                    PatchedTechData.Add(techType, techDataInstance);
                 }
-                else
-                {
-                    added++;
-                }
-
-                CraftData.techData.Add(techType, techDataInstance);
             }
+            CustomTechData.Clear();
 
             if (added > 0)
                 Logger.Log($"Added {added} new entries to the CraftData.techData dictionary.", LogLevel.Info);
 
             if (replaced > 0)
                 Logger.Log($"Replaced {replaced} existing entries to the CraftData.techData dictionary.", LogLevel.Info);
+        }
+
+        private static void PreparePrefabIDCachePostfix()
+        {
+            Dictionary<TechType, string> techMapping = CraftData.techMapping;
+            Dictionary<string, TechType> entClassTechTable = CraftData.entClassTechTable;
+            foreach (ModPrefab prefab in ModPrefab.Prefabs)
+            {
+                techMapping[prefab.TechType] = prefab.ClassID;
+                entClassTechTable[prefab.ClassID] = prefab.TechType;
+            }
         }
     }
 }
